break connection tests into multiple files
